pyproject.toml
[project]
name = "caption-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.31.0",
    "python-dotenv>=1.1.1",
]
